reorg some of the changes
