+++ src-tauri/src/managers/model.rs
@@ -4,23 +4,24 @@ use super::model_capabilities::{
 use crate::settings::{get_settings, write_settings};
 use anyhow::Result;
 use flate2::read::GzDecoder;
-use futures_util::StreamExt;
 use hf_hub::api::tokio::{ApiBuilder, CancellationToken, Progress};
 use hf_hub::{Cache, Repo, RepoType};
 use log::{debug, error, info, warn};
 use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
 use specta::Type;
 use std::collections::{HashMap, HashSet};
 use std::fs;
 use std::fs::File;
-use std::io::{Read, Write};
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
 use tar::Archive;
 use tauri::{AppHandle, Emitter, Manager};
+
+mod download;
+
+use download::{HttpDownloadOutcome, DOWNLOAD_STALL_TIMEOUT};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Type)]
 pub enum EngineType {
@@ -307,38 +308,6 @@ pub struct DownloadProgress {
     pub downloaded: u64,
     pub total: u64,
     pub percentage: f64,
-}
-
-/// Bound on connection setup for direct HTTP downloads (mirror + URL models).
-const HTTP_CONNECT_TIMEOUT: Duration = Duration::from_secs(15);
-
-/// No headers, body bytes, or hf-hub progress for this long means the transfer
-/// is wedged, not slow: direct downloads error out (keeping the partial for
-/// resume) and HF attempts are cancelled by a watchdog — either way the retry
-/// loop and mirror fallback take over instead of hanging forever.
-const DOWNLOAD_STALL_TIMEOUT: Duration = Duration::from_secs(60);
-
-/// Start offset of a `Content-Range: bytes <start>-<end>/<total>` header.
-fn content_range_start(value: &str) -> Option<u64> {
-    let range = value.trim().strip_prefix("bytes")?.trim_start();
-    range.split('-').next()?.trim().parse().ok()
-}
-
-/// How a [`ModelManager::download_http_resumable`] call ended, cancellation
-/// being an outcome (partial kept, no error surfaced) rather than a failure.
-#[derive(Debug)]
-enum HttpDownloadOutcome {
-    Completed,
-    Cancelled,
-}
-
-/// Side-channel notifications from the resumable HTTP downloader, decoupled
-/// from Tauri (the production wrapper maps them onto app events) so the
-/// transport logic is testable without an `AppHandle`.
-enum HttpDownloadEvent<'a> {
-    Progress(&'a DownloadProgress),
-    VerificationStarted,
-    VerificationCompleted,
 }
 
 /// Resolve a Hugging Face model file in the shared HF cache, if already present.
@@ -1877,56 +1846,6 @@ impl ModelManager {
         })
     }
 
-    /// Verifies the SHA256 of `path` against `expected_sha256` (if provided).
-    /// On mismatch or read error the partial file is deleted and an error is returned,
-    /// so the next download attempt always starts from a clean state.
-    /// When `expected_sha256` is `None` (custom user models) verification is skipped.
-    fn verify_sha256(path: &Path, expected_sha256: Option<&str>, model_id: &str) -> Result<()> {
-        let Some(expected) = expected_sha256 else {
-            return Ok(());
-        };
-        match Self::compute_sha256(path) {
-            Ok(actual) if actual == expected => {
-                info!("SHA256 verified for model {}", model_id);
-                Ok(())
-            }
-            Ok(actual) => {
-                warn!(
-                    "SHA256 mismatch for model {}: expected {}, got {}",
-                    model_id, expected, actual
-                );
-                let _ = fs::remove_file(path);
-                Err(anyhow::anyhow!(
-                    "Download verification failed for model {}: file is corrupt. Please retry.",
-                    model_id
-                ))
-            }
-            Err(e) => {
-                let _ = fs::remove_file(path);
-                Err(anyhow::anyhow!(
-                    "Failed to verify download for model {}: {}. Please retry.",
-                    model_id,
-                    e
-                ))
-            }
-        }
-    }
-
-    /// Computes the SHA256 hex digest of a file, reading in 64KB chunks to handle large models.
-    fn compute_sha256(path: &Path) -> Result<String> {
-        let mut file = File::open(path)?;
-        let mut hasher = Sha256::new();
-        let mut buffer = [0u8; 65536];
-        loop {
-            let n = file.read(&mut buffer)?;
-            if n == 0 {
-                break;
-            }
-            hasher.update(&buffer[..n]);
-        }
-        Ok(format!("{:x}", hasher.finalize()))
-    }
-
     /// Download a Hugging Face-sourced model into the shared HF cache via
     /// hf-hub, reporting progress through the same `model-download-progress`
     /// event the URL path uses. Uses hf-hub's stock cache, but deliberately
@@ -2229,286 +2148,6 @@ impl ModelManager {
                 Ok(true)
             }
         }
-    }
-
-    /// Emit verification events around a blocking sha256 check of `path`.
-    /// On mismatch `verify_sha256` deletes the file, so the next attempt (or
-    /// next source) starts clean. A `None` hash skips checking (custom models).
-    async fn verify_file_with_events(
-        model_id: &str,
-        path: &Path,
-        expected_sha256: Option<&str>,
-        emit: &(dyn Fn(HttpDownloadEvent<'_>) + Send + Sync),
-    ) -> Result<()> {
-        emit(HttpDownloadEvent::VerificationStarted);
-        let path = path.to_path_buf();
-        let expected = expected_sha256.map(str::to_string);
-        let id = model_id.to_string();
-        tokio::task::spawn_blocking(move || Self::verify_sha256(&path, expected.as_deref(), &id))
-            .await
-            .map_err(|e| anyhow::anyhow!("SHA256 task panicked: {}", e))??;
-        emit(HttpDownloadEvent::VerificationCompleted);
-        Ok(())
-    }
-
-    /// [`Self::download_http_resumable_with_events`] wired to the Tauri event
-    /// bus — the production entry point.
-    async fn download_http_resumable(
-        &self,
-        model_id: &str,
-        url: &str,
-        partial_path: &Path,
-        expected_size: Option<u64>,
-        expected_sha256: Option<&str>,
-        cancel_token: &CancellationToken,
-    ) -> Result<HttpDownloadOutcome> {
-        let app_handle = self.app_handle.clone();
-        let id = model_id.to_string();
-        Self::download_http_resumable_with_events(
-            model_id,
-            url,
-            partial_path,
-            expected_size,
-            expected_sha256,
-            cancel_token,
-            &move |event| {
-                let _ = match event {
-                    HttpDownloadEvent::Progress(progress) => {
-                        app_handle.emit("model-download-progress", progress)
-                    }
-                    HttpDownloadEvent::VerificationStarted => {
-                        app_handle.emit("model-verification-started", &id)
-                    }
-                    HttpDownloadEvent::VerificationCompleted => {
-                        app_handle.emit("model-verification-completed", &id)
-                    }
-                };
-            },
-        )
-        .await
-    }
-
-    /// The one resumable HTTP downloader, shared by the mirror fallback and
-    /// URL-sourced models: fetch `url` into `partial_path`, resuming what's
-    /// already there, and leave verified bytes in `partial_path` on success —
-    /// finalizing (rename / extract) is the caller's job. Takes progress and
-    /// verification notifications as a callback instead of touching Tauri, so
-    /// the failure-mode behavior below is exercised by tests against a local
-    /// socket server.
-    ///
-    /// Robustness properties, in the order the failure modes appear:
-    /// - a partial already at the expected size (crash between completion and
-    ///   finalize) is verified and accepted instead of asking the server for
-    ///   `Range: bytes=<EOF>-` and looping on 416 forever; an oversized one is
-    ///   deleted; a live 416 finishes the partial only when a hash can prove
-    ///   it, and otherwise clears it
-    /// - connection setup and every body chunk are bounded by
-    ///   [`HTTP_CONNECT_TIMEOUT`] / [`DOWNLOAD_STALL_TIMEOUT`] and race the
-    ///   cancel token, so a wedged transfer can neither hang the download
-    ///   forever nor ignore a cancel
-    /// - a 200 to a Range request (server ignored it) restarts from zero
-    ///   rather than appending the whole file to the partial; a 206 must start
-    ///   exactly at our offset or the partial is discarded
-    /// - a server claiming or sending more than the expected size is cut off
-    ///   at the first excess byte, not trusted until it closes the stream
-    /// - the final bytes are checked against `expected_size` (catalog, or
-    ///   content-length when unknown) and `expected_sha256` before returning
-    async fn download_http_resumable_with_events(
-        model_id: &str,
-        url: &str,
-        partial_path: &Path,
-        expected_size: Option<u64>,
-        expected_sha256: Option<&str>,
-        cancel_token: &CancellationToken,
-        emit: &(dyn Fn(HttpDownloadEvent<'_>) + Send + Sync),
-    ) -> Result<HttpDownloadOutcome> {
-        let mut resume_from = partial_path.metadata().map(|m| m.len()).unwrap_or(0);
-
-        if let Some(expected) = expected_size {
-            if resume_from > expected {
-                let _ = fs::remove_file(partial_path);
-                resume_from = 0;
-            } else if resume_from == expected && expected > 0 {
-                info!(
-                    "Partial download of {} is already full-size; verifying",
-                    model_id
-                );
-                Self::verify_file_with_events(model_id, partial_path, expected_sha256, emit)
-                    .await?;
-                return Ok(HttpDownloadOutcome::Completed);
-            }
-        }
-
-        if resume_from > 0 {
-            info!(
-                "Resuming download of {} from byte {}",
-                model_id, resume_from
-            );
-        } else {
-            info!("Starting fresh download of {} from {}", model_id, url);
-        }
-
-        let client = reqwest::Client::builder()
-            .connect_timeout(HTTP_CONNECT_TIMEOUT)
-            .build()?;
-        let mut request = client.get(url);
-        if resume_from > 0 {
-            request = request.header("Range", format!("bytes={}-", resume_from));
-        }
-        let response = tokio::select! {
-            r = tokio::time::timeout(DOWNLOAD_STALL_TIMEOUT, request.send()) => r
-                .map_err(|_| anyhow::anyhow!(
-                    "no response within {}s from {}",
-                    DOWNLOAD_STALL_TIMEOUT.as_secs(), url
-                ))??,
-            _ = cancel_token.cancelled() => return Ok(HttpDownloadOutcome::Cancelled),
-        };
-
-        // 416 to our Range request means its start is at or past the object's
-        // end. With a catalog size in hand that can only mean the server's
-        // object is *smaller* than expected (a full-size partial never issues
-        // a request — handled above), and with no hash there is no trusted
-        // signal to bless the partial: both restart clean. Only a hash can
-        // genuinely finish a partial here. Without a Range in flight a 416 is
-        // just a broken server, which the generic status check below rejects.
-        if resume_from > 0 && response.status() == reqwest::StatusCode::RANGE_NOT_SATISFIABLE {
-            if expected_size.is_some() || expected_sha256.is_none() {
-                let _ = fs::remove_file(partial_path);
-                return Err(anyhow::anyhow!(
-                    "server object ends before the expected size (HTTP 416)"
-                ));
-            }
-            Self::verify_file_with_events(model_id, partial_path, expected_sha256, emit).await?;
-            return Ok(HttpDownloadOutcome::Completed);
-        }
-        // A 200 to a Range request means the server ignored it and is sending
-        // the whole file; appending it to the partial would corrupt the model.
-        if resume_from > 0 && response.status() == reqwest::StatusCode::OK {
-            let _ = fs::remove_file(partial_path);
-            resume_from = 0;
-        }
-        if !response.status().is_success() {
-            return Err(anyhow::anyhow!(
-                "server returned HTTP {}",
-                response.status()
-            ));
-        }
-        // On a 206, trust but verify the offset: a reply starting anywhere but
-        // exactly our partial's end would silently corrupt the file on append.
-        if resume_from > 0 && response.status() == reqwest::StatusCode::PARTIAL_CONTENT {
-            let starts_at = response
-                .headers()
-                .get(reqwest::header::CONTENT_RANGE)
-                .and_then(|v| v.to_str().ok())
-                .and_then(content_range_start);
-            if starts_at != Some(resume_from) {
-                let _ = fs::remove_file(partial_path);
-                return Err(anyhow::anyhow!(
-                    "server returned Content-Range starting at {:?}, expected {}",
-                    starts_at,
-                    resume_from
-                ));
-            }
-        }
-        // When the catalog pins the size, a server advertising a different
-        // total is already misbehaving — reject before writing anything.
-        if let (Some(expected), Some(len)) = (expected_size, response.content_length()) {
-            if resume_from + len != expected {
-                return Err(anyhow::anyhow!(
-                    "server advertises {} bytes, expected {}",
-                    resume_from + len,
-                    expected
-                ));
-            }
-        }
-
-        let known_total =
-            expected_size.or_else(|| response.content_length().map(|l| resume_from + l));
-        let total_size = known_total.unwrap_or(0);
-        let mut downloaded = resume_from;
-        let mut file = if resume_from > 0 {
-            std::fs::OpenOptions::new()
-                .append(true)
-                .open(partial_path)?
-        } else {
-            std::fs::File::create(partial_path)?
-        };
-
-        let emit_progress = |downloaded: u64| {
-            emit(HttpDownloadEvent::Progress(&DownloadProgress {
-                model_id: model_id.to_string(),
-                downloaded,
-                total: total_size,
-                percentage: if total_size > 0 {
-                    (downloaded as f64 / total_size as f64) * 100.0
-                } else {
-                    0.0
-                },
-            }));
-        };
-        emit_progress(downloaded);
-
-        // Throttle progress events to max 10/sec (100ms intervals)
-        let mut last_emit = Instant::now();
-        let throttle = Duration::from_millis(100);
-        let mut stream = response.bytes_stream();
-        loop {
-            let chunk = tokio::select! {
-                c = tokio::time::timeout(DOWNLOAD_STALL_TIMEOUT, stream.next()) => match c {
-                    // Stalled mid-body: keep the partial for resume.
-                    Err(_) => return Err(anyhow::anyhow!(
-                        "transfer stalled: no data for {}s",
-                        DOWNLOAD_STALL_TIMEOUT.as_secs()
-                    )),
-                    Ok(None) => break,
-                    Ok(Some(chunk)) => chunk?,
-                },
-                _ = cancel_token.cancelled() => {
-                    // Keep the partial for resume; caller handles state cleanup.
-                    return Ok(HttpDownloadOutcome::Cancelled);
-                }
-            };
-            // An untrusted server must not be able to fill the disk: cut the
-            // transfer at the first byte past the known total instead of
-            // trusting it to eventually close the stream. Everything written
-            // so far is tainted by a provably-misbehaving server — clear it.
-            if let Some(cap) = known_total {
-                if downloaded + chunk.len() as u64 > cap {
-                    drop(file);
-                    let _ = fs::remove_file(partial_path);
-                    return Err(anyhow::anyhow!(
-                        "server sent more than the expected {} bytes",
-                        cap
-                    ));
-                }
-            }
-            file.write_all(&chunk)?;
-            downloaded += chunk.len() as u64;
-            if last_emit.elapsed() >= throttle {
-                emit_progress(downloaded);
-                last_emit = Instant::now();
-            }
-        }
-        file.flush()?;
-        drop(file);
-        emit_progress(downloaded);
-
-        if let Some(expected) = known_total {
-            let actual = partial_path.metadata()?.len();
-            if actual != expected {
-                let _ = fs::remove_file(partial_path);
-                return Err(anyhow::anyhow!(
-                    "download incomplete: expected {} bytes, got {}",
-                    expected,
-                    actual
-                ));
-            }
-        }
-
-        // The catalog hash is the trust anchor: for a mirror (an untrusted
-        // host) this verification is what makes the fallback safe at all.
-        Self::verify_file_with_events(model_id, partial_path, expected_sha256, emit).await?;
-        Ok(HttpDownloadOutcome::Completed)
     }
 
     pub async fn download_model(&self, model_id: &str) -> Result<()> {
@@ -3240,75 +2879,6 @@ mod tests {
         assert!(!models.contains_key(default_filename.trim_end_matches(".gguf")));
     }
 
-    // ── SHA256 verification tests ─────────────────────────────────────────────
-
-    /// Helper: write `data` to a temp file and return (TempDir, path).
-    /// TempDir must be kept alive for the duration of the test.
-    fn write_temp_file(data: &[u8]) -> (TempDir, std::path::PathBuf) {
-        let dir = TempDir::new().unwrap();
-        let path = dir.path().join("model.partial");
-        let mut f = File::create(&path).unwrap();
-        f.write_all(data).unwrap();
-        (dir, path)
-    }
-
-    #[test]
-    fn test_verify_sha256_skipped_when_none() {
-        // Custom models have no expected hash — verification must be a no-op.
-        let (_dir, path) = write_temp_file(b"anything");
-        assert!(ModelManager::verify_sha256(&path, None, "custom").is_ok());
-        assert!(
-            path.exists(),
-            "file must be untouched when verification is skipped"
-        );
-    }
-
-    #[test]
-    fn test_verify_sha256_passes_on_correct_hash() {
-        // Compute the real hash so the test is self-consistent.
-        let (_dir, path) = write_temp_file(b"hello world");
-        let actual = ModelManager::compute_sha256(&path).unwrap();
-        assert!(
-            ModelManager::verify_sha256(&path, Some(&actual), "test_model").is_ok(),
-            "should pass when hash matches"
-        );
-        assert!(
-            path.exists(),
-            "file must be kept on successful verification"
-        );
-    }
-
-    #[test]
-    fn test_verify_sha256_fails_and_deletes_partial_on_mismatch() {
-        let (_dir, path) = write_temp_file(b"this is not the real model");
-        let wrong_hash = "0000000000000000000000000000000000000000000000000000000000000000";
-
-        let result = ModelManager::verify_sha256(&path, Some(wrong_hash), "bad_model");
-
-        assert!(result.is_err(), "mismatch must return an error");
-        assert!(
-            result.unwrap_err().to_string().contains("corrupt"),
-            "error message should mention corruption"
-        );
-        assert!(
-            !path.exists(),
-            "partial file must be deleted after hash mismatch"
-        );
-    }
-
-    #[test]
-    fn test_verify_sha256_fails_and_deletes_partial_when_file_missing() {
-        // Simulate a partial file that was already removed (e.g. disk full mid-download).
-        let dir = TempDir::new().unwrap();
-        let missing_path = dir.path().join("gone.partial");
-        // Don't create the file — it should not exist.
-
-        let result =
-            ModelManager::verify_sha256(&missing_path, Some("anyexpectedhash"), "missing_model");
-
-        assert!(result.is_err(), "missing file must return an error");
-    }
-
     fn push_gguf_str(out: &mut Vec<u8>, val: &str) {
         out.extend_from_slice(&(val.len() as u64).to_le_bytes());
         out.extend_from_slice(val.as_bytes());
@@ -3382,509 +2952,6 @@ mod tests {
         assert!(
             !models.contains_key("someone/llama-7b/llama-q8.gguf"),
             "non-ASR gguf must be ignored"
-        );
-    }
-
-    // ── Resumable HTTP downloader tests ───────────────────────────────────────
-    //
-    // Each test drives `download_http_resumable_with_events` against a scripted
-    // single-connection server on a local socket — no Tauri, no real network.
-    // Success means verified bytes are left in the partial (finalizing is the
-    // caller's job), so "no Completed on a bad hash" is the rename gate too.
-
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    use tokio::net::{TcpListener, TcpStream};
-
-    fn sha_hex(data: &[u8]) -> String {
-        format!("{:x}", Sha256::digest(data))
-    }
-
-    fn http_response(status_line: &str, headers: &[String], body: &[u8]) -> Vec<u8> {
-        let mut head = format!("HTTP/1.1 {}\r\nConnection: close\r\n", status_line);
-        for h in headers {
-            head.push_str(h);
-            head.push_str("\r\n");
-        }
-        head.push_str("\r\n");
-        let mut bytes = head.into_bytes();
-        bytes.extend_from_slice(body);
-        bytes
-    }
-
-    async fn read_request_head(sock: &mut TcpStream) -> String {
-        let mut buf = Vec::new();
-        let mut byte = [0u8; 1];
-        while !buf.ends_with(b"\r\n\r\n") {
-            if sock.read_exact(&mut byte).await.is_err() {
-                break;
-            }
-            buf.push(byte[0]);
-        }
-        String::from_utf8_lossy(&buf).into_owned()
-    }
-
-    /// Serve exactly one connection: capture the request head, write
-    /// `response` verbatim, close. Returns the URL to fetch and a handle
-    /// yielding the captured request head (lowercased for header asserts).
-    async fn serve_once(response: Vec<u8>) -> (String, tokio::task::JoinHandle<String>) {
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let handle = tokio::spawn(async move {
-            let (mut sock, _) = listener.accept().await.unwrap();
-            let head = read_request_head(&mut sock).await;
-            // The client may hang up early (error tests); that's not a
-            // server-side failure.
-            let _ = sock.write_all(&response).await;
-            let _ = sock.shutdown().await;
-            head.to_lowercase()
-        });
-        (format!("http://{}/file", addr), handle)
-    }
-
-    async fn run_download(
-        url: &str,
-        partial: &Path,
-        expected_size: Option<u64>,
-        expected_sha256: Option<&str>,
-        cancel: &CancellationToken,
-    ) -> Result<HttpDownloadOutcome> {
-        ModelManager::download_http_resumable_with_events(
-            "test/model",
-            url,
-            partial,
-            expected_size,
-            expected_sha256,
-            cancel,
-            &|_| {},
-        )
-        .await
-    }
-
-    #[tokio::test]
-    async fn http_fresh_download_completes_and_verifies() {
-        let body = b"hello world";
-        let (url, server) = serve_once(http_response(
-            "200 OK",
-            &[format!("Content-Length: {}", body.len())],
-            body,
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-
-        let out = run_download(
-            &url,
-            &partial,
-            Some(body.len() as u64),
-            Some(&sha_hex(body)),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Completed));
-        assert_eq!(fs::read(&partial).unwrap(), body);
-        let head = server.await.unwrap();
-        assert!(
-            !head.contains("range:"),
-            "fresh download must not send a Range header"
-        );
-    }
-
-    #[tokio::test]
-    async fn http_resume_appends_with_valid_content_range() {
-        let full = b"helloworld";
-        let (url, server) = serve_once(http_response(
-            "206 Partial Content",
-            &[
-                "Content-Range: bytes 5-9/10".to_string(),
-                "Content-Length: 5".to_string(),
-            ],
-            &full[5..],
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, &full[..5]).unwrap();
-
-        let out = run_download(
-            &url,
-            &partial,
-            Some(10),
-            Some(&sha_hex(full)),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Completed));
-        assert_eq!(fs::read(&partial).unwrap(), full);
-        let head = server.await.unwrap();
-        assert!(head.contains("range: bytes=5-"), "got request: {head}");
-    }
-
-    #[tokio::test]
-    async fn http_wrong_content_range_discards_partial() {
-        // Server answers 206 but restarts the range at 0: appending would
-        // corrupt, so the partial must be dropped and the call must fail.
-        let full = b"helloworld";
-        let (url, _server) = serve_once(http_response(
-            "206 Partial Content",
-            &[
-                "Content-Range: bytes 0-9/10".to_string(),
-                "Content-Length: 10".to_string(),
-            ],
-            full,
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, &full[..5]).unwrap();
-
-        let err = run_download(
-            &url,
-            &partial,
-            Some(10),
-            Some(&sha_hex(full)),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(err.to_string().contains("Content-Range"), "{err}");
-        assert!(!partial.exists(), "corrupt-prone partial must be deleted");
-    }
-
-    #[tokio::test]
-    async fn http_range_ignored_200_restarts_from_zero() {
-        let body = b"helloworld!";
-        let (url, _server) = serve_once(http_response(
-            "200 OK",
-            &[format!("Content-Length: {}", body.len())],
-            body,
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, b"XXXXX").unwrap(); // stale bytes that must not survive
-
-        let out = run_download(
-            &url,
-            &partial,
-            Some(body.len() as u64),
-            Some(&sha_hex(body)),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Completed));
-        assert_eq!(fs::read(&partial).unwrap(), body);
-    }
-
-    #[tokio::test]
-    async fn http_416_with_hash_finalizes_complete_partial() {
-        // URL-model shape: no catalog size, but a hash — the one case where a
-        // 416 may bless the partial, because verification proves it.
-        let body = b"the whole file";
-        let (url, _server) = serve_once(http_response("416 Range Not Satisfiable", &[], b"")).await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, body).unwrap();
-
-        let out = run_download(
-            &url,
-            &partial,
-            None,
-            Some(&sha_hex(body)),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Completed));
-        assert_eq!(fs::read(&partial).unwrap(), body);
-    }
-
-    #[tokio::test]
-    async fn http_416_with_wrong_hash_clears_partial() {
-        let (url, _server) = serve_once(http_response("416 Range Not Satisfiable", &[], b"")).await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, b"corrupted bytes").unwrap();
-
-        let err = run_download(
-            &url,
-            &partial,
-            None,
-            Some(&sha_hex(b"the real file")),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(err.to_string().contains("corrupt"), "{err}");
-        assert!(
-            !partial.exists(),
-            "failed verification must delete the partial"
-        );
-    }
-
-    #[tokio::test]
-    async fn http_416_short_of_expected_size_clears_partial() {
-        // Mirror shape: catalog size known, partial shorter, yet the server
-        // says our offset is past EOF — its object is smaller than the catalog
-        // expects. Never blessed, even with a hash on hand.
-        let (url, _server) = serve_once(http_response("416 Range Not Satisfiable", &[], b"")).await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, b"12345").unwrap();
-
-        let err = run_download(
-            &url,
-            &partial,
-            Some(10),
-            Some(&sha_hex(b"1234567890")),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(err.to_string().contains("416"), "{err}");
-        assert!(!partial.exists());
-    }
-
-    #[tokio::test]
-    async fn http_416_without_trust_signals_clears_partial() {
-        // No size, no hash: nothing can prove the partial complete, so a 416
-        // must restart clean instead of accepting unverified bytes.
-        let (url, _server) = serve_once(http_response("416 Range Not Satisfiable", &[], b"")).await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, b"whatever").unwrap();
-
-        let err = run_download(&url, &partial, None, None, &CancellationToken::new())
-            .await
-            .unwrap_err();
-
-        assert!(err.to_string().contains("416"), "{err}");
-        assert!(!partial.exists());
-    }
-
-    #[tokio::test]
-    async fn http_full_size_partial_verified_without_network() {
-        // Crash-between-completion-and-finalize recovery: the partial already
-        // has every byte, so no request is issued at all (the URL points at a
-        // dead port to prove it).
-        let body = b"complete content";
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, body).unwrap();
-
-        let out = run_download(
-            "http://127.0.0.1:9/unreachable",
-            &partial,
-            Some(body.len() as u64),
-            Some(&sha_hex(body)),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Completed));
-        assert_eq!(fs::read(&partial).unwrap(), body);
-    }
-
-    #[tokio::test]
-    async fn http_oversized_partial_restarts_clean() {
-        let body = b"12345";
-        let (url, server) = serve_once(http_response(
-            "200 OK",
-            &[format!("Content-Length: {}", body.len())],
-            body,
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-        fs::write(&partial, b"garbage-beyond-expected").unwrap();
-
-        let out = run_download(
-            &url,
-            &partial,
-            Some(body.len() as u64),
-            Some(&sha_hex(body)),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Completed));
-        assert_eq!(fs::read(&partial).unwrap(), body);
-        let head = server.await.unwrap();
-        assert!(
-            !head.contains("range:"),
-            "oversized partial must restart, not resume: {head}"
-        );
-    }
-
-    #[tokio::test]
-    async fn http_conflicting_content_length_rejected_before_writing() {
-        let (url, _server) = serve_once(http_response(
-            "200 OK",
-            &["Content-Length: 10".to_string()],
-            b"0123456789",
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-
-        let err = run_download(
-            &url,
-            &partial,
-            Some(5),
-            Some(&sha_hex(b"01234")),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(err.to_string().contains("advertises"), "{err}");
-        assert!(
-            !partial.exists(),
-            "nothing may be written on an up-front size conflict"
-        );
-    }
-
-    #[tokio::test]
-    async fn http_body_exceeding_expected_size_aborts_and_clears() {
-        // Chunked response (no Content-Length to reject up-front) that keeps
-        // sending past the expected size: the disk-fill guard must cut the
-        // transfer at the first excess byte and clear the tainted partial.
-        let (url, _server) = serve_once(http_response(
-            "200 OK",
-            &["Transfer-Encoding: chunked".to_string()],
-            b"3\r\nabc\r\n5\r\ndefgh\r\n0\r\n\r\n",
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-
-        let err = run_download(
-            &url,
-            &partial,
-            Some(5),
-            Some(&sha_hex(b"abcde")),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(err.to_string().contains("more than"), "{err}");
-        assert!(!partial.exists(), "tainted partial must be deleted");
-    }
-
-    #[tokio::test]
-    async fn http_wrong_hash_after_download_clears_partial() {
-        let body = b"tampered bytes";
-        let (url, _server) = serve_once(http_response(
-            "200 OK",
-            &[format!("Content-Length: {}", body.len())],
-            body,
-        ))
-        .await;
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-
-        let err = run_download(
-            &url,
-            &partial,
-            Some(body.len() as u64),
-            Some(&sha_hex(b"the expected bytes")),
-            &CancellationToken::new(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(err.to_string().contains("corrupt"), "{err}");
-        assert!(!partial.exists());
-    }
-
-    #[tokio::test]
-    async fn http_cancel_while_awaiting_headers() {
-        // Server accepts and goes silent. Cancellation must win immediately —
-        // not after the stall timeout, and certainly not never.
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        tokio::spawn(async move {
-            let (mut sock, _) = listener.accept().await.unwrap();
-            let _ = read_request_head(&mut sock).await;
-            tokio::time::sleep(Duration::from_secs(600)).await;
-        });
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-
-        let cancel = CancellationToken::new();
-        let trigger = cancel.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_millis(100)).await;
-            trigger.cancel();
-        });
-
-        let started = Instant::now();
-        let out = run_download(
-            &format!("http://{}/file", addr),
-            &partial,
-            Some(10),
-            Some("00"),
-            &cancel,
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Cancelled));
-        assert!(
-            started.elapsed() < Duration::from_secs(30),
-            "cancel must not wait out the stall timeout"
-        );
-    }
-
-    #[tokio::test]
-    async fn http_cancel_during_stalled_body_keeps_partial() {
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        tokio::spawn(async move {
-            let (mut sock, _) = listener.accept().await.unwrap();
-            let _ = read_request_head(&mut sock).await;
-            let _ = sock
-                .write_all(b"HTTP/1.1 200 OK\r\nContent-Length: 100\r\n\r\nabc")
-                .await;
-            tokio::time::sleep(Duration::from_secs(600)).await;
-        });
-        let dir = TempDir::new().unwrap();
-        let partial = dir.path().join("m.partial");
-
-        let cancel = CancellationToken::new();
-        let trigger = cancel.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_millis(300)).await;
-            trigger.cancel();
-        });
-
-        let out = run_download(
-            &format!("http://{}/file", addr),
-            &partial,
-            Some(100),
-            Some("00"),
-            &cancel,
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(out, HttpDownloadOutcome::Cancelled));
-        assert_eq!(
-            fs::read(&partial).unwrap(),
-            b"abc",
-            "bytes received before the cancel must be kept for resume"
         );
     }
 }
